backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.6.4
python-socketio[asgi]==5.11.2
aiofiles==23.2.1
python-dateutil==2.9.0